move gates into ops folder for later adding measurement, classical control
