+++ qutip/qip/algorithms/qft.py
@@ -39,7 +39,7 @@ __all__ = ['qft', 'qft_steps', 'qft_gate_sequence']
 import numpy as np
 import scipy.sparse as sp
 from qutip.qobj import *
-from qutip.qip.gates import snot, cphase, swap
+from qutip.qip.ops import snot, cphase, swap
 from qutip.qip.circuit import QubitCircuit
