+++ qutip/qip/circuit.py
@@ -42,7 +42,7 @@ import numpy as np
 import numpy as np
 
 from qutip.qip.circuit_latex import _latex_compile
-from qutip.qip.gates import *
+from qutip.qip.ops import *
 from qutip.qip.qubits import qubit_states
 
 __all__ = ['Gate', 'QubitCircuit']
