+++ qutip/qip/ops/gates.py
