+++ qutip/tests/test_qpt.py
@@ -36,7 +36,7 @@ from numpy.testing import assert_, run_module_suite
 import scipy.linalg as la
 
 from qutip import (spre, spost, qeye, sigmax, sigmay, sigmaz, qpt)
-from qutip.qip.gates import snot, cnot
+from qutip.qip.ops import snot, cnot
 
 
 def test_qpt_snot():
